Changed Compi_Options_Table Class
The Compi_Options_Table class only provides the structure. Instantiating
the class, filling the stack with a new object then getting its values
into a new attribute is redundant. I removed the constructor and made a
static get_dash_options() method that returns the structure as an array.
I hate static too but it's simpler this way. :D